refactor(03-03): rewire consumers to relocated enums; delete string->enum maps (M2-07)
- event.py: import FillStatus from core.enums; replace fill_status_map.get + buggy ValueError with FillStatus(status)
- transaction.py: replace transaction_type_map.get + ('Value %s', x) ValueError with TransactionType(action)
- Add _missing_ to TransactionType (Rule 3) so uppercase action strings parse case-insensitively
- cash/position/metrics/transaction managers: import relocated enums from core.enums; delete local defs + dead 'from enum import Enum'
- Repoint order_manager + portfolio_handler FillStatus import to core.enums (Rule 3: mypy strict explicit-export)
- FillStatus.EXECUTED -> OrderStatus.FILLED reconciliation preserved; behavioral oracle byte-exact
